PriceWise/backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
dnspython==2.7.0
dotenv==0.9.9
ecdsa==0.19.1
email-validator==2.2.0
fastapi==0.116.1
greenlet==3.2.3
h11==0.16.0
httptools==0.6.4
idna==3.10
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyodbc==5.2.0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.1
typing_extensions==4.14.1
typing-inspection==0.4.1
uvicorn[standard]==0.35.0
watchfiles==1.1.0
websockets==15.0.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.131.0

